feat(validation): update NotConflictAccountValidator to use Long for count feat(config): update application.yml comments for clarity and change server port to 8085 test(service): add unit tests for SysUserServiceImpl with create, update, query, and delete functionalities

+++ src/main/java/com/bosyon/zisnackdesk/validation/NotConflictAccountValidator.java
@@ -26,7 +26,7 @@ public class NotConflictAccountValidator implements ConstraintValidator<NotConfl
 
         LambdaQueryWrapper<SysUser> wrapper = new LambdaQueryWrapper<>();
         wrapper.eq(SysUser::getAccount, account).isNull(SysUser::getDeletedAt);
-        Integer count = sysUserMapper.selectCount(wrapper);
+        Long count = sysUserMapper.selectCount(wrapper);
         return count == null || count == 0;
     }
 }
